% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack_improve.R
\name{brute_force_knapsack_improve}
\alias{brute_force_knapsack_improve}
\title{Brute force approach with profile and optimize the code.}
\usage{
brute_force_knapsack_improve(x, W)
}
\arguments{
\item{x}{data frame with two columns \code{w} (weight) and \code{v} (value) of items to place in the knapsack}

\item{W}{the maximum weight (numeric) the knapsack can hold}
}
\value{
theoretical maximum \code{$value} (knapsack value) composed of \code{$elements} (which items)
}
\description{
Brute force approach with profile and optimize the code.
}
