% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack_parallel.R
\name{brute_force_knapsack_parallel}
\alias{brute_force_knapsack_parallel}
\title{Brute force approach for the knapsack problem with parallel.}
\usage{
brute_force_knapsack_parallel(x, W, parallel = TRUE, fast = FALSE)
}
\arguments{
\item{x}{data frame with two columns \code{w} (weight) and \code{v} (value) of items to place in the knapsack #'}

\item{W}{the maximum weight (numeric) the knapsack can hold}

\item{parallel}{parallel}

\item{fast}{whether to use the rcpp optimization}
}
\value{
theoretical maximum \code{$value} (knapsack value) composed of \code{$elements} (which items) #'
}
\description{
Brute force approach for the knapsack problem with parallel.
}
