% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cppfor}
\alias{cppfor}
\title{Fast Implementation of Dynamic Algorithm Using C++}
\arguments{
\item{n}{The number of available objects to place in the knapsack}

\item{w}{A numeric vector which contains the weights from data.frame 'x'}

\item{v}{A numeric vector which contains the values from data.frame 'x'}

\item{mat}{A zero matrix with (n + 1) rows and (W + 1) columns, with n
being the number of available objects and W being the weight limit of the
knapsack}

\item{mem}{A zero data.frame with one column and (n + 1) rows}
}
\description{
This function creates an equivalent of the nested for loop in
the dynamic algorithm for a faster implementation using C++ code instead of
R code.
}
