% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_knapsack.R
\name{create_knapsack}
\alias{create_knapsack}
\title{Generates a knapsack object}
\usage{
create_knapsack(n)
}
\arguments{
\item{n}{number of elements in the knapsack object (numeric)}
}
\value{
data frame with two columns \code{w} (weight) and \code{v} (value)
}
\description{
Generates a knapsack object
}
