% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_knapsack_data.R
\name{create_knapsack_data}
\alias{create_knapsack_data}
\title{Create knapsack}
\usage{
create_knapsack_data(
  n,
  seed = set.seed(42, kind = "Mersenne-Twister", normal.kind = "Inversion")
)
}
\arguments{
\item{n}{is the number of samples}

\item{seed}{is defines which seed to use. Default value is 42 with some extra specifications of kind.}
}
\value{
returns a data.frame of generated knapsack data. The data has 2 columns and n rows.
}
\description{
Create knapsack
}
