% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_knapsack.R
\name{dynamic_knapsack_r}
\alias{dynamic_knapsack_r}
\title{R implementation of dynamic programming for knapsack problem}
\usage{
dynamic_knapsack_r(x, W)
}
\arguments{
\item{x}{Dataframe}

\item{W}{Numeric scalar}
}
\value{
List with two elements:
        1. value - optimum value that can be included in the knapsack
        2. elements - the list of elements that can be included in the knapsack
}
\description{
R implementation of dynamic programming for knapsack problem
}
