% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dynamic_optimized.R
\name{Dynamic_optimized_v1}
\alias{Dynamic_optimized_v1}
\title{Optimized Dynamic Programming}
\usage{
Dynamic_optimized_v1(x, W)
}
\arguments{
\item{x}{data.frame where each row represents an item with its weight and value.}

\item{W}{A numeric value representing maximum capacity of knapsack.}
}
\value{
final A list shows the maximum knapsack value and the elements.
}
\description{
This function solves the knapsack problem using an optimized dynamic 
 programming approach.It returns the maximum value that can be achieved 
 within the weight limit and the indices of the selected items.
}
\examples{
 items <- data.frame(weight = c(2, 3, 4), value = c(3, 4, 5))
 W <- 5
 Dynamic_optimized_v1(items, W)


}
