% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_knapsack.R
\name{generate_knapsack}
\alias{generate_knapsack}
\title{generate_knapsack}
\usage{
generate_knapsack(n = 2000, seed = 42)
}
\arguments{
\item{n}{number of samples to generate.}

\item{seed}{which seed to use.}
}
\value{
data.frame with the columns w and v.
}
\description{
\code{generate_knapsack} generates a dataset for example usage of the functions.
}
\examples{
generate_knapsack()

}
