% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_knapsack_objects.R
\name{get_knapsack_objects}
\alias{get_knapsack_objects}
\title{get_knapsack_objects}
\usage{
get_knapsack_objects(n)
}
\arguments{
\item{n}{number of items that should be created}
}
\value{
Dataframe with items' weights and values
}
\description{
Get a dataframe to use for the knapsack problem.
It contains two columns: w = weights, v = values
}
\examples{
get_knapsack_objects(2000)

}
