% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Solving the knapsack problem through heuristic approach}
\usage{
greedy_knapsack(x, W, opt = TRUE)
}
\arguments{
\item{x}{is a data.frame containing knapsack data. It has two columns: v with values and w with weights.}

\item{W}{is a numeric scalar which defined the maximum possible weight knapsack can handle before breaking.}

\item{opt}{indicates whether or not the better version should be used (for tests it is set to FALSE)}
}
\value{
a list object containing the best bag value given the weight and the elements that make up the best bag.
}
\description{
Solving the knapsack problem through heuristic approach
}
