% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack_c.R
\name{greedy_knapsack_C}
\alias{greedy_knapsack_C}
\title{greedy_knapsack_c function}
\usage{
greedy_knapsack_C(x, W)
}
\arguments{
\item{x}{A data.frame with two variables v (values) and w (weights)}

\item{W}{The maximum knapsack size}
}
\value{
A list of the maximum value in the knapsack and the elements choosen to be in the knapsack
}
\description{
greedy_knapsack_c function
}
