% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_optimized.R
\name{greedy_knapsack_v2}
\alias{greedy_knapsack_v2}
\title{Optimized Greedy Algorithm}
\usage{
greedy_knapsack_v2(x, W)
}
\arguments{
\item{x}{data.frame data.frame where each row represents an item with its 
weight and value.}

\item{W}{A numeric value representing maximum capacity of knapsack.}
}
\value{
final A list shows the maximum knapsack value and the elements.
}
\description{
This function solves the knapsack problem using an optimized greedy algorithm 
 by reducing Unnecessary Variables and matrices.
}
\examples{
items <- data.frame(weight = c(2, 3, 4), value = c(3, 4, 5))
W <- 5
greedy_knapsack_v2(items, W)

}
