% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{intToBinary}
\alias{intToBinary}
\title{The knapsack problem: brute force algorithm.}
\usage{
intToBinary(x, head)
}
\arguments{
\item{x}{an object of class data.frame with two variables v (values) and w (weights).}

\item{head}{numeric scalar object that represents the knapsack size.}
}
\value{
\code{brute_force_knapsack} returns a list with two elements: the elements added to the knapsack and the maximum knapsack value.
}
\description{
The knapsack problem: brute force algorithm.
}
\examples{
intToBinary(4,8)


}
