% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack.R
\docType{package}
\name{knapsack-package}
\alias{knapsack}
\alias{knapsack-package}
\title{knapsack:}
\description{
Implements three algorithms for solving the knapsack problem.
}
\author{
\strong{Maintainer}: Theodor Emanuelsson \email{theodor.emanuelsson@me.com}

}
