% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack.R
\docType{class}
\name{knapsack_class}
\alias{knapsack_class}
\title{Knapsack Algorithms}
\arguments{
\item{ks_dataset}{is a data.frame generated by initialzed function which contain variables W and V, weights and values}

\item{ks_size}{is the total size of the knapsack}
}
\value{
This function will return a list of selected elements and their total values

This function will return a list of selected elements and their total values

This function will return a list of selected elements and their total values
}
\description{
This class will generate a dataset for knapsack in initialized function
}
\details{
Bruteforce, dynamimc programminf and greedy algorithm for the knapsack problem

This algorithm gives all possbile values with good accuracy, and also gives the maximum value for the knapsack

This algorithm gives all possbile values with good accuracy, and also gives the maximum value for the knapsack

This algorithm gives all possbile values with good accuracy, and also gives the maximum value for the knapsack
}

\examples{
\dontrun{
obj <- knapsack_class$new()
}
\dontrun{
obj$knapsack_brute_force(obj$ks_dataset[1:8,],3500)
obj$knapsack_brute_force(obj$ks_dataset[1:12,],3500)
obj$knapsack_brute_force(obj$ks_dataset[1:8,],2000)
}
\dontrun{
obj$dynamic_programming(obj$ks_dataset[1:8,],3500)
obj$dynamic_programming(obj$ks_dataset[1:12,],3500)
obj$dynamic_programming(obj$ks_dataset[1:8,],2000)
}
\dontrun{
obj$greedy_heuristich(obj$ks_dataset[1:8,],3500)
obj$greedy_heuristich(obj$ks_dataset[1:12,],3500)
obj$greedy_heuristich(obj$ks_dataset[1:8,],2000)
}
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem}
}
\author{
Mahmood Siddique
}
