% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Solving the knapsack problem through dynamic programming}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{is a data.frame containing knapsack data. It has two columns: v with values and w with weights.}

\item{W}{is a numeric scalar which defined the maximum possible weight knapsack can handle before breaking.}
}
\value{
a list object containing the best bag value given the weight and the elements that make up the best bag.
}
\description{
Solving the knapsack problem through dynamic programming
}
