% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_dynamic.R
\name{knapsack_dynamic_cpp}
\alias{knapsack_dynamic_cpp}
\title{Dynammic Programming for Knapsack Problem with RCPP}
\usage{
knapsack_dynamic_cpp(x, W)
}
\arguments{
\item{x}{}

\item{W}{}
}
\value{
It returns the maximum value obtained and selected items.
}
\description{
The algorithm solves the knapsack problem using Dynammic Programming approach with RCPP.
}
\examples{
x <- data.frame(v = c(10, 5, 15, 7, 6, 18), w = c(2, 3, 5, 7, 1, 4))
x <- as.matrix(x)
W <- 15
knapsack_dynamic_cpp(x,W)
}
\seealso{
[Knapsack Problem](https://en.wikipedia.org/wiki/Knapsack_problem#0.2F1_knapsack_problem)
}
