% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_knapsack_improve.R
\name{knapsack_dynamic_improve}
\alias{knapsack_dynamic_improve}
\title{Dynamic approach for the knapsack problem with improvement}
\usage{
knapsack_dynamic_improve(x, W)
}
\arguments{
\item{x}{data frame with two columns \code{w} (weight) and \code{v} (value) of items to place in the knapsack}

\item{W}{the maximum weight (numeric) the knapsack can hold}
}
\value{
theoretical maximum \code{$value} (knapsack value) composed of \code{$elements} (which items)
}
\description{
Dynamic approach for the knapsack problem with improvement
}
