% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_objects.R
\docType{data}
\name{knapsack_objects}
\alias{knapsack_objects}
\title{Knapsack Objects Data}
\format{
A \code{dataframe} of dimensions 2000x2 containing weights \code{w} and values\code{v}.
}
\usage{
data(knapsack_objects)
}
\description{
Knapsack Objects Data
}
\examples{
data(knapsack_objects)
}
\keyword{datasets}
