% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{knapSackdynamic_cpp}
\alias{knapSackdynamic_cpp}
\title{Rcc method implementation for knapsack dynamic programming}
\usage{
knapSackdynamic_cpp(MaxWeight, weights, value, length)
}
\arguments{
\item{MaxWeight}{numeric vector indicating maximum capacity of knapsack}

\item{weights}{numeric vector indicating weight of individual element in the datase}

\item{value}{value of individual element in the dataset}

\item{length}{number of objects in the data set}
}
\value{
a matrix with sequence of all weights and values.
}
\description{
Rcc method implementation for knapsack dynamic programming
}
