% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{object}
\alias{object}
\title{Generate Knapsack Objects}
\usage{
object(n)
}
\arguments{

\item{n}{An integer specifying the number of objects to generate.}

}
\value{
knapsack_objects A data frame containing the generated knapsack 
  objects with columns for weight (`w`) and value (`v`).
}
\description{
This function generates a data frame of knapsack objects with random weights
 and values.
}
