% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/old_brute_force_knapsack.R
\name{old_brute_force_knapsack}
\alias{old_brute_force_knapsack}
\title{Brute force search}
\usage{
old_brute_force_knapsack(x, W)
}
\arguments{
\item{x}{a dataframe consists of two variables \code{w} and \code{v}
\itemize{
 \item w: weight of each element in the knapsack
 \item v: value of each element in the knapsack
}}

\item{W}{a numeric value of the knapsack size.}
}
\value{
A list of maximum knapsack value and which elements that can add to the knapsack.
}
\description{
It is the only real solution that is guaranteed to give a correct answer in all situations for the knapsack problem.
Going through all the possible alternatives choices and in the end the algorithm return the maximum value.
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem}
}

