% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimized_brute_force.R
\name{optimized_brute_force}
\alias{optimized_brute_force}
\title{optimized_brute_force}
\usage{
optimized_brute_force(x, W)
}
\arguments{
\item{x}{Dataframe containing weight and value of objects}

\item{W}{maximum weight allowed in knapsack}
}
\value{
List with the optimal value and the respective elements,
that are included in the solution.
}
\description{
Optimized version of the Brute Force Algorithm
}
\examples{
knapsack_objects <- get_knapsack_objects(2000)

optimized_brute_force(x = knapsack_objects[1:8, ], W = 3500)

}
