% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimized_dynamic_programming.R
\name{optimized_dynamic_programming}
\alias{optimized_dynamic_programming}
\title{optimized_dynamic_programming}
\usage{
optimized_dynamic_programming(x, W)
}
\arguments{
\item{x}{Dataframe containing weight and value of objects}

\item{W}{maximum weight allowed in knapsack}
}
\value{
List with the optimal value and the respective elements,
that are included in the solution.
}
\description{
Optimized version of the dynamic programming algorithm.
}
\examples{
knapsack_objects <- get_knapsack_objects(2000)

optimized_dynamic_programming(x = knapsack_objects[1:8, ], W = 3500)

}
