% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_sac_a_dos}
\alias{rcpp_sac_a_dos}
\title{Algorithme Brut Force using C++}
\usage{
rcpp_sac_a_dos(poids, valeurs, poids_max)
}
\arguments{
\item{poids}{vecteur contenant les poids des objets}

\item{valeurs}{vecteur contenant les valeurs (numériques) des objets}

\item{poids_max}{représente le poids maximal autorisé dans le sac}
}
\value{
La solution optimale
}
\description{
Algorithme Brut Force using C++
}
