% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_sac_a_dos_dp}
\alias{rcpp_sac_a_dos_dp}
\title{Algorithme Dynamic Programmation using C++}
\usage{
rcpp_sac_a_dos_dp(poids, valeurs, poids_max, precision = 1000L)
}
\arguments{
\item{poids}{vecteur contenant les poids des objets}

\item{valeurs}{vecteur contenant les valeurs (numériques) des objets}

\item{poids_max}{représente le poids maximal autorisé dans le sac}

\item{precision}{represente la precision}
}
\value{
La solution optimale
}
\description{
Algorithme Dynamic Programmation using C++
}
