% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_sac_a_dos_glouton}
\alias{rcpp_sac_a_dos_glouton}
\title{Algorithme Glouton using C++}
\usage{
rcpp_sac_a_dos_glouton(poids, valeurs, poids_max)
}
\arguments{
\item{poids}{Vecteur contenant les poids des objets}

\item{valeurs}{Vecteur contenant les valeurs (numériques) des objets}

\item{poids_max}{Entier représentant le poids maximal autorisé dans le sac}
}
\value{
Une solution... optimale ?
}
\description{
Algorithme Glouton using C++
}
