% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sac_a_dos.R
\name{sac_a_dos}
\alias{sac_a_dos}
\title{Sac a dos : Algorithme Brut Force}
\usage{
sac_a_dos(poids, valeurs, poids_max)
}
\arguments{
\item{poids}{vecteur contenant les poids des objets}

\item{valeurs}{vecteur contenant les valeurs (numériques) des objets}

\item{poids_max}{représente le poids maximal autorisé dans le sac}
}
\value{
La solution optimale
}
\description{
Algorithme naif testant toutes les combinaisons possibles
}
