% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sac_a_dos_dp.R
\name{sac_a_dos_dp}
\alias{sac_a_dos_dp}
\title{Sac a dos Prog Dyn}
\usage{
sac_a_dos_dp(poids, valeurs, poids_max)
}
\arguments{
\item{poids}{vecteur de poids}

\item{valeurs}{vecteur des valeurs}

\item{poids_max}{représente le poids maximal autorisé}
}
\value{
Une solution optimale ?
}
\description{
Algorithme Programmation Dynamique
}
