% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brute_force_search.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute Force method to solve knapsack problem}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{Data Frama having weights(w) and values(v) columns}

\item{W}{Positive number weight of knapsack}

\item{parallel}{Argument to parallelize brute force}
}
\value{
The maximum value and elements added to Knapsack
}
\description{
Brute Force method to solve knapsack problem
}
