% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_brute_force.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Knapsack brute force algorithm implementation}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{Data.frame with variables 'v' and 'w'}

\item{W}{Maximum weight that knapsack can be packed with}
}
\value{
list with elements 
  \item 'elements' denoting which objects in x that were used
  \item 'value' denoting the sum of value for the used objects
}
\description{
Knapsack brute force algorithm implementation
}
