% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Using brute force search for the knapsack problem}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{Data frame with two variables v and w}

\item{W}{knapsack size}
}
\value{
maximum knapsack value and which elements
}
\description{
Using brute force search for the knapsack problem
}
