% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute Force Solution for Knapsack Problem}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{Dataset as a dataframe}

\item{W}{Kanapsack size}
}
\value{
The knapsack with the largest value of the elements added to the knapsack
}
\description{
Brute Force Solution for Knapsack Problem
}
\examples{
brute_force_knapsack(x = knapsack_objects[1:8,], W = 3500)
brute_force_knapsack(x = knapsack_objects[1:12,], W = 3500)
brute_force_knapsack(x = knapsack_objects[1:8,], W = 2000)
brute_force_knapsack(x = knapsack_objects[1:12,], W = 2000)

}
