% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute Force Solution for Knapsack Problem}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{A data.frame with two columns: w (the weights of the objects) and
v (the values of the objects)}

\item{W}{The maximum weight the knapsack can hold}
}
\value{
A named list with two elements: value (which displays the highest
value that can be held in the knapsack) and elements (a vector listing the
item numbers in said knapsack)
}
\description{
Brute Force Solution for Knapsack Problem
}
\examples{
brute_force_knapsack(x = knapsack_objects[1:8,], W = 3500)
brute_force_knapsack(x = knapsack_objects[1:12,], W = 3500)
brute_force_knapsack(x = knapsack_objects[1:8,], W = 2000)
brute_force_knapsack(x = knapsack_objects[1:12,], W = 2000)

}
