% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brute_force.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute force knapsack}
\usage{
brute_force_knapsack(x, W, fast = FALSE)
}
\arguments{
\item{x}{A data.frame containing with variables \code{w} and \code{v}}

\item{W}{A numeric value stating the maximum allowed weight of selected objects}
}
\value{
A list containg the best value and the elements selected
}
\description{
A brute force approach for solving knapsack problems. The function will not work for more than 32 items.
}
\examples{
n <- 2000
knapsack_objects <- data.frame(w=sample(1:4000, size = n, replace = TRUE), v=runif(n = n, 0, 10000))
brute_force_knapsack(x = knapsack_objects[1:8,], W = 3500)
brute_force_knapsack(x = knapsack_objects[1:12,], W = 3500)
brute_force_knapsack(x = knapsack_objects[1:8,], W = 2000)
brute_force_knapsack(x = knapsack_objects[1:12,], W = 2000)
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem}
}
