% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute force search function}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{A data.frame with two variables v (values) and w (weights)}

\item{W}{The maximum knapsack weight}
}
\value{
The maximum value in the knapsack and the elements choosen to be in the knapsack
}
\description{
Brute force search function
}
