% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute force search for knapsack problem}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{is object that contains two vectors with weights and values. Must be a data frame with two colums w and v.}

\item{W}{is the maximum capacity(weight) of the knapsack and should be a positive numeric number.}
}
\value{
A list with two elements, maximum value and seleted items each.
}
\description{
The only solution that is guaranteed to give a correct answer in all situations for the knapsack problem is using brute-force search, i.e. going through all possible alternatives and return the maximum value found.
}
