% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R,
%   R/brute_force_knapsack_parallel.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Optimized brute force search for knapsack problem}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)

brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{An object of data.frame with 2 columns with colnames w and v.}

\item{W}{A single positive value.}

\item{parallel}{If TRUE parallelize the function across detected cores, default is FALSE.}
}
\value{
The function returns a list containing the maximum knapsack value and which elements.

The function returns a list containing the maximum knapsack value and which elements.
}
\description{
Optimized brute force search for knapsack problem

Optimized brute force search for knapsack problem
}
\examples{
set.seed(42)
 n <- 2000
knapsack_objects <-
 data.frame(
   w=sample(1:4000, size = n, replace = TRUE),
   v=runif(n = n, 0, 10000)
 ) 
results <- brute_force_knapsack(x = knapsack_objects[1:8,], W=3500)
set.seed(42)
 n <- 2000
knapsack_objects <-
 data.frame(
   w=sample(1:4000, size = n, replace = TRUE),
   v=runif(n = n, 0, 10000)
 ) 
results <- brute_force_knapsack(x = knapsack_objects[1:8,], W=3500)
}
