% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Solve the 0-1 knapsack problem (brute force approach)}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{A dataframe with weights (w) and values (v).}

\item{W}{the knapsack size.}
}
\value{
a list with the maximum value and the indexes of the elements that compose this optimal solution
}
\description{
Solve the 0-1 knapsack problem (brute force approach)
}
