% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Title}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{data.frame cx with two variables v and w}

\item{W}{Weight}

\item{parallel}{default false}
}
\value{
final
}
\description{
Title
}
