% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute Force Knapsack Algorithm
The function returns maximum value within all the possible solutions}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{data.frame, contains variables weight(w) and value(v)}

\item{W}{weight threshold for the knapsack}
}
\value{
Total and Maximum value within the threshold of W
}
\description{
Brute Force Knapsack Algorithm
The function returns maximum value within all the possible solutions
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack problem}
}
