% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute force algorithm for the knapsack problem}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{is a data.frame containing variables 'w' and 'v', weights and values}

\item{W}{is the total size of the knapsack}
}
\description{
Brute force algorithm for the knapsack problem
}
\details{
This algorithm gives all possible values with good accuracy , and also gives the maximum value for the knapsack
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem}
}
\seealso{
\code{\link{knapsack_dynamic}}
\code{\link{greedy_knapsack}}
}
\author{
Mahnaz Mohammadzamani
}
