% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute force.}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{data.frame}

\item{W}{integer}
}
\value{
list of ideal value and elements that solves the problem.
}
\description{
Solving Knapsack problem with brute force approach.
}
