% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{brute_force:knapsack}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{a data frame with two variables, v (value) and w (weight).}

\item{W}{the knapsack size}
}
\value{
The function returns a list with two elements: the value of the knapsack and the items it consists of.
}
\description{
A function that solves the knapsack problem:
we have a knapsack that can take a limited weight W, which we want to fill with a number of items i = 1,...,n, each with
a weight w_i and a value v_i. The goal is to find the knapsack with the largest value of the elements added to the knapsack.
}
\examples{
set.seed(42)
n <- 2000
knapsack_objects<-data.frame(w =sample(1: 4000,size= n,replace= TRUE),
                            v =runif(n = n,0, 10000))

brute_force_knapsack(x = knapsack_objects [1: 12,],W = 3500)
brute_force_knapsack(x = knapsack_objects [1:12,],W = 2000)
}
