% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute Force Algorithm}
\usage{
brute_force_knapsack(x, W, parallel)
}
\arguments{
\item{x}{Dataframe}

\item{W}{Numeric scalar}

\item{parallel}{Logical scalar}
}
\value{
List with two elements:
        1. value - optimum value that can be included in the knapsack
        2. elements - the list of elements that can be included in the knapsack
}
\description{
This function finds the optimum value for the knapsack problem.
             Parallelization can be achieved by setting parallel as TRUE
}
\details{
Advanced Programming in R - Lab 6
Knapsack Problem - Brute Force Algorithm. This approach finds the exact
optimimum solution by comparing each of the possible outcomes and has high
computational complexity. Parallelization has been implemented for this
technique to improve the speed of the algorithm
}
\examples{
brute_force_knapsack(x = knapsack_objects[1:12, ], W = 3500)
}
