% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute force knapsack}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{Knapsack data frame}

\item{W}{Knapsack size}
}
\value{
elements and the sum of the corresponding values
}
\description{
Brute force knapsack
}
