% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{brute_force_knapsack}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{Dataframe containing weight and value of objects}

\item{W}{maximum weight allowed in knapsack}

\item{parallel}{boolean value whether to use parallelization}
}
\value{
List with the optimal value and the respective elements,
that are included in the solution.
}
\description{
Solver for the knapsack problem based on the brute force algorithm.
}
\examples{
knapsack_objects <- get_knapsack_objects(2000)

brute_force_knapsack(x = knapsack_objects[1:8, ], W = 3500)

}
