% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Solve the Knapsack Problem using Brute Force (with optional parallelization)}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{A data frame with two columns: \code{v} for values and \code{w} for weights.}

\item{W}{A positive numeric value representing the maximum weight capacity of the knapsack.}

\item{parallel}{A logical value indicating whether to parallelize the search. Defaults to FALSE.}
}
\value{
A list with two elements:
  \item{value}{The maximum total value of the knapsack.}
  \item{elements}{The indices of the items in \code{x} that make up the optimal solution.}
}
\description{
This function solves the knapsack problem using brute-force search.
It optionally parallelizes the brute-force search if the argument \code{parallel} is set to TRUE.
}
