% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Knapsack Brute Force}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{a data.frame with two variables v (Value) and w (Weight) and returns the maximum knapsack value and which elements (rows in the data.frame).}

\item{W}{is the knapsack size.}

\item{parallel}{Default: False. Use paralleling to solve problem}
}
\value{
returns the maximum knapsack value and which elements (rows in the data.frame).
}
\description{
Knapsack Brute Force
}
\examples{
x <- data.frame(v = c(10, 5, 15, 7, 6, 18), w = c(2, 3, 5, 7, 1, 4))
W <- 15
brute_force_knapsack(x,W)
}
