% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute Force}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{x \code{data.frame} containing objects with weights \code{w} and values \code{v} as columns.}

\item{W}{max weight of the knapsack}

\item{parallel}{boolean indicating whether or not parallel computing should be used}
}
\value{
a list with max value and corresponding elements
}
\description{
Brute Force
}
\examples{
data(knapsack_objects)
brute_force_knapsack(x = knapsack_objects[1:8,], W = 3500)
}
