% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute Force}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{}

\item{W}{}
}
\value{
a list with max value and corresponding elements
}
\description{
Brute Force
}
\examples{
knapsack_objects <-
data.frame(
  v = c(10,40,30,50),
  w = c(5,4,6,3)
)
W <- 10

brute_force_knapsack((knapsack_objects, W)
}
