% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute force solve for the knapsack problem}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{A data frame with two variables w = weight and v = value}

\item{W}{A numeric scalar with the maximum weight the knapsack is allowed to contain}
}
\value{
The function returns a list with the optimum packed value and the corresponding elements
}
\description{
A function that uses brute force to solve the knapsack problem
}
\examples{
n <- 10
knapsack_objects <- data.frame(w = sample(1:4000, size = n, replace = TRUE),
                            v = runif(n = n, 0, 10000))
brute_force_knapsack(x = knapsack_objects, W = 3500)

}
