% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute force search for knapsack problem}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{data frame with two columns \code{w} (weight) and \code{v} (value) of items to place in the knapsack}

\item{W}{the maximum weight (numeric) the knapsack can hold}
}
\value{
theoretical maximum \code{$value} (knapsack value) composed of \code{$elements} (which items)
}
\description{
Brute force search for knapsack problem
}
