% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_algos.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute-force knapsack solver}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{data.frame with columns w (weights, positive integers) and v (values, positive numerics)}

\item{W}{numeric capacity (positive)}

\item{parallel}{logical; if TRUE attempt to use multiple cores (detects available cores)}
}
\value{
list(value = numeric total value, elements = integer vector of original indices chosen)
}
\description{
Brute-force knapsack solver
}
\examples{
x <- data.frame(w = c(2,3,4), v = c(3,4,5)); brute_force_knapsack(x, 5)
}
