% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Solving the knapsack problem using brute force}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{an data frame with a number of items.}

\item{W}{the limited weight the knapsack can handle.}
}
\value{
A list of the maximated value with corresponding elements that fits in the bag.
}
\description{
The function solves the knapsack problem , \url{https://en.wikipedia.org/wiki/Knapsack_problem}
using brute force search.
}
\examples{
n<- 2000
knapsack_objects <-data.frame(w=sample(1:4000, size = n, replace = TRUE),v=runif(n = n, 0, 10000))
brute_force_knapsack(x=knapsack_objects[1:8,], W=2000)
}

