% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack_parallel.R
\name{brute_force_knapsack_parallel}
\alias{brute_force_knapsack_parallel}
\title{Brute Force Knapsack Algorithm
The function returns maximum value within all the possible solutions}
\usage{
brute_force_knapsack_parallel(x, W, parallel = FALSE)
}
\arguments{
\item{x}{data.frame, contains variables weight(w) and value(v)}

\item{W}{weight threshold for the knapsack}

\item{parallel}{Boolean, if 'TRUE' then implements palatalization}
}
\value{
Total and Maximum value within the threshold of W
\url{https://en.wikipedia.org/wiki/Knapsack problem}
}
\description{
Brute Force Knapsack Algorithm
The function returns maximum value within all the possible solutions
}
\seealso{
\code{\link{greedy_knapsack}}
\code{\link{knapsack_dynamic}}
}
