% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_programming_knapsack.R
\name{dynamic_knapsack}
\alias{dynamic_knapsack}
\title{Dynamic programming method to solve knapsack problem}
\usage{
dynamic_knapsack(x, W)
}
\arguments{
\item{x}{Data Frama having weights(w) and values(v) columns}

\item{W}{Positive number weight of knapsack}
}
\value{
The maximum value and elements added to Knapsack
}
\description{
Dynamic programming method to solve knapsack problem
}
