% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Knapsack.R
\name{dynamic_knapsack}
\alias{dynamic_knapsack}
\title{Solve the 0-1 knapsack problem (dynamic progeamming approach)}
\usage{
dynamic_knapsack(x, W)
}
\arguments{
\item{x}{A dataframe with weights (w) and values (v).}

\item{W}{the knapsack size.}
}
\value{
a list with the maximum value and the indexes of the elements that compose this optimal solution
}
\description{
Solve the 0-1 knapsack problem (dynamic progeamming approach)
}
