% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_knapsack.R
\name{dynamic_knapsack}
\alias{dynamic_knapsack}
\title{Dynamic Programming Algorithm}
\source{
/url{"https://en.wikipedia.org/wiki/Knapsack_problem#0.2F1_knapsack_problem"}
}
\usage{
dynamic_knapsack(x, W)
}
\arguments{
\item{x}{Dataframe}

\item{W}{Numeric scalar}
}
\value{
List with two elements:
        1. value - optimum value that can be included in the knapsack
        2. elements - the list of elements that can be included in the knapsack
}
\description{
This function finds the optimum value for the knapsack problem.
}
\details{
Advanced Programming in R - Lab 6
Knapsack Problem - Dynamic Programming Algorithm. This approach finds the exact
optimimum solution by building an optimization matrix and has lower complexity
than the brute force algorithm. The speed of this algorithm is further improved
by implementing it in Rcpp. (Refer function dynamic_knapsack_cpp)
}
\examples{
dynamic_knapsack(x = knapsack_objects[1:12, ], W = 3500)
}
