% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic.R
\name{dynamic_knapsack}
\alias{dynamic_knapsack}
\title{Solve the Knapsack Problem using Dynamic Programming}
\usage{
dynamic_knapsack(x, W)
}
\arguments{
\item{x}{A data frame with two columns: \code{v} for values and \code{w} for weights.
Each row represents an item with a value and a weight.}

\item{W}{A positive numeric value representing the maximum weight capacity of the knapsack.}
}
\value{
A list with two elements:
  \item{value}{The maximum total value of the knapsack.}
  \item{elements}{The indices of the items in \code{x} that make up the optimal solution.}
}
\description{
This function solves the knapsack problem using dynamic programming.
It constructs a table to store the maximum value for each weight capacity.
The result is obtained by iterating over all items and weights.
}
\examples{
knapsack_objects <- data.frame(
  w = sample(1:4000, size = 8, replace = TRUE),
  v = runif(n = 8, 0, 10000)
)
dynamic_knapsack(knapsack_objects, W = 3500)

}
