% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_method.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy Heuristic method to solve knapsack problem}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{Data Frama having weights(w) and values(v) columns}

\item{W}{Positive number weight of knapsack}
}
\value{
The maximum value and elements added to Knapsack
}
\description{
Greedy Heuristic method to solve knapsack problem
}
