% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_greedy_heuristic.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Knapsack greedy algorithm implementation}
\usage{
greedy_knapsack(x, W, fast = FALSE)
}
\arguments{
\item{x}{Data.frame with variables 'v' and 'w'}

\item{W}{Maximum weight that knapsack can be packed with}

\item{fast}{not implemented}
}
\value{
list with elements 
\itemize{
  \item {'elements' denoting which objects in x that were used}
  \item {'value' denoting the sum of value for the used objects}
  }
}
\description{
Knapsack greedy algorithm implementation
}
\examples{
\dontrun{
set.seed(666)
n <- 9000
stuff <- data.frame(w = sample(1:10, n, replace = TRUE),
                    v = sample(4:12, n, replace = TRUE))
# returns value = 46 made up from elements 2,4,5,7,8,9
greedy_knapsack(stuff,25)
}
}
