% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy Heuristic}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{a dataframe consists of two variables \code{w} and \code{v}
\itemize{
 \item w: weight of each element in the knapsack
 \item v: value of each element in the knapsack
}}

\item{W}{a numeric value of the knapsack size.}
}
\value{
A list of maximum knapsack value and which elements that can add to the knapsack.
}
\description{
It is one approach to solve the knapsack problem which was proposed by George Dantzig.
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem#Greedy_approximation_algorithm}
}

