% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Greedy_heuristic.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy heuristic knapsack}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{A data.frame containing with variables \code{w} and \code{v}}

\item{W}{A numeric value stating the maximum allowed weight of selected objects}
}
\value{
A list containg the best value and the elements selected
}
\description{
A greedy heuristic for solving knapsack problems.
}
\examples{
n <- 2000
knapsack_objects <- data.frame(w=sample(1:4000, size = n, replace = TRUE), v=runif(n = n, 0, 10000))
greedy_knapsack(x = knapsack_objects[1:800,], W = 3500)
greedy_knapsack(x = knapsack_objects[1:1200,], W = 2000)
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack problem#Greedy approximation algorithm}
}
