% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{greedy_knapsack function}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{A data.frame with two variables v (values) and w (weights)}

\item{W}{The maximum knapsack size}
}
\value{
A list of the maximum value in the knapsack and the elements choosen to be in the knapsack
}
\description{
greedy_knapsack function
}
