% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{greedy_knapsack_c function}
\usage{
greedy_knapsack(x, W, fast = FALSE)
}
\arguments{
\item{x}{A data.frame with two variables v (values) and w (weights)}

\item{W}{The maximum knapsack size}

\item{fast}{=FALSE set to TRUE if you want c++ code}
}
\value{
A list of the maximum value in the knapsack and the elements choosen to be in the knapsack
}
\description{
greedy_knapsack_c function
}
