% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy heuristic approach for knapsack problem}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{is object that contains two vectors with weights and values. Must be a data frame with two colums w and v.}

\item{W}{is the maximum capacity(weight) of the knapsack and should be a positive numeric number.}
}
\value{
A list with two elements, maximum value and seleted items each.
}
\description{
This algorithm will not give an exact result (but it can be shown that it will return at least 50% of the true maximum value), but it will reduce the computational complexity considerably (actually to O(n log n) due to the sorting part of the algorithm).
}
