% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{The knapsack problem: Greedy approximation algorithm.}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{an object of class data.frame with two variables v (values) and w (weights).}

\item{W}{numeric scalar object that represents the knapsack size.}
}
\value{
\code{greedy_knapsack} returns a list with two elements: the elements added to the knapsack and the maximum knapsack value.
}
\description{
The knapsack problem is a discrete optimization problem where we have a knapsack that can take a 
limited weight W and we want to fill this knapsack with a number of items i = 1,...,n, each with 
a weight w(i) and a value v(i). The goal is to find the knapsack with the largest value of the 
elements added to the knapsack.
}
\details{
\code{greedy_knapsack} uses the Greedy approximation algorithm. This algorithm works by:

(i) Computing the value-to-weight ratios.

(ii) Sorting the items in decreasing order of the value-to-weight ratios

(iii) Placing the sorted items in the knapsack until the current weight is higher than the knapsack capacity W.
}
\examples{
knapsack_objects <- generate_knapsack()
greedy_knapsack(x = knapsack_objects[1:8,], W = 3500)
greedy_knapsack(x = knapsack_objects[1:12,], W = 3500)
greedy_knapsack(x = knapsack_objects[1:8,], W = 2000)
greedy_knapsack(x = knapsack_objects[1:12,], W = 2000)

}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem#Greedy_approximation_algorithm}
}
