% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Knapsack problem solver with greedy algorithm}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{data frame of weights (first column) and value (second column)}

\item{W}{Maximum capacity (numeric)}
}
\value{
Return the maximum value and elements (index of values) and weight
}
\description{
Knapsack problem solver with greedy algorithm
}
\examples{
set.seed(42)
n <- 2000
knapsack_objects <- data.frame(w = sample(1:4000, size = n, replace = TRUE),
    v <- runif(n = n, 0, 10000))
greedy_knapsack(x = knapsack_objects[1:800,], W = 3500)
}
