% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Title}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{data.frame cx with two variables v and w}

\item{W}{Weight}
}
\value{
final
}
\description{
Title
}
