% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_heuristic.r
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Implements Greedy ALgorithm for knapsack problem}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{data.frame, contains variables weight(w) and value(v)}

\item{W}{weight threshold for the knapsack}
}
\value{

}
\description{
Implements Greedy ALgorithm for knapsack problem
}
\examples{
greedy_knapsack(x = knapsack_objects[1:800,], W = 3500)
greedy_knapsack(x = knapsack_objects[1:1200,], W = 2000)
}
\seealso{
\code{\link{brute_force_knapsack}}
\code{\link{knapsack_dynamic}}
}
