% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy}
\usage{
greedy_knapsack(x, W, fast = FALSE)
}
\arguments{
\item{x}{data.frame}

\item{W}{integer}

\item{fast}{logical}
}
\value{
list of ideal value and elements that solves the problem
}
\description{
Solving Knapsack problem with Greedy heuristic
}
