% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{greedy_knapsack}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{a data frame with two variables, v (value) and w (weight).}

\item{W}{the knapsack size}
}
\value{
Returns a list with two elements: the value of the knapsack and the items it consists of.
}
\description{
The function gives an approximation for the problem by using George Dantzigs greedy approximation algorithm.
}
\examples{
n = 2000
knapsack_objects<-data.frame(w =sample(1: 4000,size= n,replace= TRUE),v =runif(n = n,0, 10000))
brute_force_knapsack(x = knapsack_objects [1: 12,],W = 3500)
brute_force_knapsack(x = knapsack_objects [1:12,],W = 2000)
}
