% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy Heuristics Algorithm}
\source{
\url{"https://en.wikipedia.org/wiki/Knapsack_problem#Greedy_approximation_algorithm"}
}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{Dataframe}

\item{W}{Numeric scalar}
}
\value{
List with two elements:
        1. value - optimum value that can be included in the knapsack
        2. elements - the list of elements that can be included in the knapsack
}
\description{
This function finds the optimum value for the knapsack problem.
}
\details{
Advanced Programming in R - Lab 6
Knapsack Problem - Greedy heuristics approach. This is an approximation
approach that trades the accuracy of optimization in return for a significant
gain in speed of computation.
}
\examples{
greedy_knapsack(x = knapsack_objects[1:12, ], W = 3500)
}
