% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy knapsack algorithm}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{A dataframe with columns weight (w) and value (v) of the objects}

\item{W}{A positive scalar with the capacity of the bag}
}
\value{
A list containing the \code{value} of the bag content and a vector of the \code{elements} inside
}
\description{
Implementation of the knapsack algorithm using the greedy approach.
}
\examples{
set.seed(42)
n <- 2000
knapsack_objects <- data.frame( w=sample(1:4000, size = n, replace = TRUE), v=runif(n = n, 0, 10000))
greedy_knapsack(x = knapsack_objects[1:8,], W = 3500)

}
\references{
Knapsack Problem - \url{https://en.wikipedia.org/wiki/Knapsack_problem}
}

