% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy knapsack}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{Knapsack data frame}

\item{W}{Knapsack size}
}
\value{
elements and the sum of the corresponding values
}
\description{
Greedy knapsack
}
