% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_heuristic.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{greedy_heuristic}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{Dataframe containing weight and value of objects}

\item{W}{maximum weight allowed in knapsack}
}
\value{
List with the optimal value and the respective elements,
that are included in the solution.
}
\description{
Solver for the knapsack problem based on the greedy heuristic.
}
\examples{
knapsack_objects <- get_knapsack_objects(2000)

greedy_knapsack(x = knapsack_objects[1:800,], W = 3500)
}
