% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greed.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Solve the Knapsack Problem using Greedy Heuristic}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{A data frame with two columns: \code{v} for values and \code{w} for weights.
Each row represents an item with a value and a weight.}

\item{W}{A positive numeric value representing the maximum weight capacity of the knapsack.}
}
\value{
A list with two elements:
  \item{value}{The total value of the items selected for the knapsack.}
  \item{elements}{The indices of the items in \code{x} that have been selected as numeric values.}
}
\description{
This function solves the knapsack problem using a greedy approach.
It selects items based on their value-to-weight ratio until the capacity is full.
}
\examples{
set.seed(42)
knapsack_objects <- data.frame(
  w = sample(1:4000, size = 1200, replace = TRUE),
  v = runif(1200, 0, 10000)
)
greedy_knapsack(x = knapsack_objects, W = 2000)

}
