% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy Heuristic for Knapsack Problem}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{}

\item{W}{}
}
\value{
It returns the maximum value obtained and selected items.
}
\description{
The algorithm solves the knapsack problem using Greedy approach. Its computational 
complexity is O(n*log(n))
}
\examples{
x <- data.frame(v = c(10, 5, 15, 7, 6, 18), w = c(2, 3, 5, 7, 1, 4))
W <- 15
greedy_knapsack(x,W)
}
\seealso{
[Knapsack Problem](https://en.wikipedia.org/wiki/Knapsack_problem#Greedy_approximation_algorithm)
}
