% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy Heuristic}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{}

\item{W}{}
}
\value{
a list with max value and corresponding elements
}
\description{
Greedy Heuristic
}
\examples{
knapsack_objects <-
data.frame(
  v = c(10,40,30,50),
  w = c(5,4,6,3)
)
W <- 10

greedy_knapsack((knapsack_objects, W)
}
