% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy Heuristic}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{\code{data.frame} containing objects with weights \code{w} and values \code{v} as columns.}

\item{W}{max weight of the knapsack}
}
\value{
a list with at least 50 % of the max value and corresponding elements
}
\description{
Greedy Heuristic
}
\examples{
data(knapsack_objects)
greedy_knapsack(x = knapsack_objects[1:8,], W = 3500)
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem#Greedy_approximation_algorithm}
}
