% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy Heuristic}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{}

\item{W}{}
}
\value{
a list with at least 50 % of the max value and corresponding elements
}
\description{
Greedy Heuristic
}
\examples{
data(knapsack_objects)
greedy_knapsack(x = knapsack_objects[1:800,], W = 3500)
}
