% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy Heuristic approach for the knapsack problem}
\usage{
greedy_knapsack(x, W, fast = FALSE)
}
\arguments{
\item{x}{data frame with two columns \code{w} (weight) and \code{v} (value) of items to place in the knapsack}

\item{W}{the maximum weight (numeric) the knapsack can hold}

\item{fast}{whether to use the rcpp optimization}
}
\value{
theoretical maximum \code{$value} (knapsack value) composed of \code{$elements} (which items)
}
\description{
Greedy Heuristic approach for the knapsack problem
}
