% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_algos.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy knapsack heuristic}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{data.frame with columns w, v}

\item{W}{numeric capacity}
}
\value{
list(value = numeric, elements = integer vector of original indices chosen)
}
\description{
Note: greedy is approximate; it returns a feasible solution quickly.
}
\examples{
x <- data.frame(w = c(2,3,4), v = c(3,4,5)); greedy_knapsack(x, 5)
}
