% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Solving the knapsack problem using the greedy heuristic function}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{an data frame with a number of items.}

\item{W}{the limited weight the knapsack can handle.}
}
\value{
A list of the maximated value with corresponding elements that fits in the bag.
}
\description{
The function solves the knapsack problem using greedy heuristic function. It won't return the exact answer however, it will reduce the complexity of the computation. It will return at least 50% of the maximum value.
}
\examples{
n<- 2000
knapsack_objects <-data.frame(w=sample(1:4000, size = n, replace = TRUE),v=runif(n = n, 0, 10000))
greedy_knapsack(x=knapsack_objects[1:8,], W=2000)

}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem#Greedy_approximation_algorithm}
}

