\name{knapsack-package}
\alias{knapsack-package}
\alias{knapsack}
\docType{package}
\title{
  Knapsack Problem
}
\description{
  Ce package propose plusieurs algorithmes pour résoudre le problème classique du sac à dos,
  notamment une méthode exhaustive (force brute), une méthode gloutonne et une approche par programmation dynamique.
  Des versions optimisées en C++ sont fournies via Rcpp pour comparer les performances.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Alexandre Combeau, Mohtadi Hammami, Ninette Hounkponou
}