\name{KnapSack-package}
\alias{KnapSack-package}
\alias{KnapSack}
\docType{package}
\title{
\packageTitle{KnapSack}
}
\description{
\packageDescription{KnapSack}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{KnapSack}
\packageIndices{KnapSack}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{KnapSack}

Maintainer: \packageMaintainer{KnapSack}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
