\name{knapsack-package}
\alias{knapsack-package}
\alias{knapsack}
\docType{package}
\title{package
\packageTitle{knapsack}
}
\description{
\packageDescription{knapsack}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{knapsack}
\packageIndices{knapsack}
%%~~ An overview of how to use the package, including the most important ~~
%%~~ functions ~~
}
\author{
\packageAuthor{knapsack}

Maintainer: \packageMaintainer{knapsack}
}
%%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%%~~ the R documentation directory ~~
\keyword{ package }

