\name{knapsack-package}
\alias{knapsack-package}
\docType{package}
\title{
Knapsack
}
\description{
Contain tree different functions to solve the knapsack problem.
}

\details{
\tabular{ll}{
Package: \tab knapsack \cr
Type: \tab Package \cr
Version: \tab 1.0\cr
Date: \tab 2015-10-01\cr
}}



\author{
Martina Sandberg <marsa505@student.liu.se>

Caroline Svahn <carsv733@student.liu.se>
}

\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem}
}

\keyword{knapsack}


