\name{Knapsack-package}
\alias{Knapsack-package}
\alias{Knapsack}
\docType{package}
\title{
\packageTitle{Knapsack}
}
\description{
\packageDescription{Knapsack}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Knapsack}
\packageIndices{Knapsack}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Knapsack}

Maintainer: \packageMaintainer{Knapsack}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
