\name{knapsack-package}
\alias{knapsack-package}
\docType{package}
\title{
  Knapsack Problems
}
\description{
  Solves knapsack, change making, subset sum, special assignment, and
  bin packing problems.

  Can handle the classical 01 single knapsack problem for real and integer 
  numbers. Provides exact and approximate solutions for large problem sets
  and for bounded and unbounded problem formulations. Certain classes of 
  multiple knapsack and assignment problems will be solved exactly as well
  as approximately.
}
\details{
\tabular{ll}{
Package: \tab knapsack\cr
Type: \tab Package\cr
Version: \tab 0.2.0\cr
Date: \tab 2012-02-02\cr
License: \tab Free for research purposes and academic use.
}
}
\author{
  Hans W Borchers for the R part. The Fortran programs have beeen written
  by Silvano Martello and Paolo Toth for their book ``Knapsack Problems''
  and are provided here with permission.

  Maintainer: HwB <hwborchers@googlemail.com>
}
\references{
  Martello, S., and P. Toth (1990). ``Knapsack Problems: Algorithms and
  Computer Implementations''. John Wiley & Sons, Ltd.
}
\keyword{ package }
\seealso{
  Other packages with knapsack solutions.
}
\examples{
##  See examples on the function help pages.
}
