\name{knapsack-package}
\alias{knapsack-package}
\alias{knapsack}
\docType{package}
\title{
\packageTitle{knapsack}
}
\description{
\packageDescription{knapsack}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{knapsack}
\packageIndices{knapsack}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{knapsack}

Maintainer: \packageMaintainer{knapsack}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
