\name{knapsack-package}
\alias{knapsack-package}
\alias{knapsack}
\docType{package}
\title{
\packageTitle{knapsack}
}
\description{
\packageDescription{knapsack}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{knapsack}
\packageIndices{knapsack}
An overview of how to use the package, including the most important functions
}
\author{
\packageAuthor{knapsack}

Maintainer: \packageMaintainer{knapsack}
}
\references{
Literature or other references for background information
}

\keyword{ package }

