\name{knapSack-package}
\alias{knapSack-package}
\alias{knapSack}
\docType{package}
\title{
\packageTitle{knapSack}
}
\description{
\packageDescription{knapSack}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{knapSack}
\packageIndices{knapSack}
}
\author{
\packageAuthor{knapSack}
Maintainer: \packageMaintainer{knapSack}
}
\keyword{ package }

