% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack.R
\docType{package}
\name{knapsack}
\alias{knapsack}
\title{The Knapsack Problem}
\description{
A package that illustrates computational complexity by comparing
diffent methods of solving the Knapsack Problem.
}
\author{
Connor Turner and Sachini Bambaranda
}
