% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack.R
\docType{package}
\name{knapsack}
\alias{knapsack}
\title{knapsack: solving the knapsack problem}
\description{
This package provides three different functions to solve the knapsack problem:
\itemize{
  \item \code{\link{brute_force_knapsack}}
  \item \code{\link{knapsack_dynamic}}
  \item \code{\link{greedy_knapsack}}
}
}
