% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{knapsack}
\alias{knapsack}
\alias{knapsack-package}
\title{knapsack: The knapsack problem.}
\description{
This package contains three different functions that are three different approaches (of different complexity) to solve the knapsack problem.
}
\author{
Laura Julia Melis and Martin Svensson.

Course in Advanced Programming in R.

Department of Computer and Information Science, Linköping University.
}
