% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{knapsack}
\alias{knapsack}
\title{the knapsack package}
\description{
The package contains different functions for solving what is called the knapsack problem. The knapsack problem is a discrete optimization problem where we have a knapsack that can take a limited weight W and we want to fill this knapsack with a number of items i = 1, ..., n, each with a weight w[i] and a value v[i]. The goal is to find the knapsack with the largest value of the elements added to the knapsack.
}
\author{
Tea Unnebäck and Karl Lindberg \email{tea@folkdatorn.se}
}
