% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_brute_force.r
\name{knapsack_brute_force}
\alias{knapsack_brute_force}
\title{Brute-Force Solution for the Knapsack Problem}
\usage{
knapsack_brute_force(x, W, parallel = FALSE)
}
\arguments{
\item{x}{a \code{data.frame} with two columns, "w" and "v", indicating the weight and value of items respectively}

\item{W}{the maximum weight that the knapsack can hold}

\item{parallel}{TRUE if a parallelised version of the solution is to be called; FALSE by default}
}
\value{
a \code{list} containing two elements: \code{value} and \code{elements}, representing the maximum value that can be realised with the given knapsack and the indices of the items to be packed in the knapsack for realising maximum value
}
\description{
Brute-Force Solution for the Knapsack Problem
}
