% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{0/1 knapsack problem}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{a dataframe consists of two variables \code{w} and \code{v}. Both variables contain only positive values.
\itemize{
 \item w: weight of each element in the knapsack
 \item v: value of each element in the knapsack
}}

\item{W}{a numeric value of the knapsack size.}
}
\value{
A list of maximum knapsack value and which elements that can add to the knapsack.
}
\description{
It is one approach to solve the knapsack problem by using Dynamic-Programming.
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem#0.2F1_knapsack_problem}
}

