% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{A Dynamic Programming Model for the Knapsack Problem}
\usage{
knapsack_dynamic(x, W, fast = FALSE)
}
\arguments{
\item{x}{A data.frame with two columns: w (the weights of the objects) and
v (the values of the objects)}

\item{W}{The maximum weight the knapsack can hold}

\item{fast}{If TRUE, utilizes C++ code to run the function faster. Default
is FALSE}
}
\value{
A named list with two elements: value (which displays the highest
value that can be held in the knapsack) and elements (a vector listing the
item numbers in said knapsack)
}
\description{
A Dynamic Programming Model for the Knapsack Problem
}
\examples{
knapsack_dynamic(x = knapsack_objects[1:8,], W = 3500)

}
