% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Finds the optimal solution for given knapsack problem}
\usage{
knapsack_dynamic(x, W, fast = FALSE)
}
\arguments{
\item{x, W, fast}{Takes in two arguments, the first one is a dataframe which contains the weights and values vectors and second one is the maximum weight that a knapsack can hold.}
}
\value{
Returns the highest value which satisfying the given weight constraint.
}
\description{
Finds the optimal solution for given knapsack problem
}
\examples{
\dontrun{ 
set.seed(42)
n <- 2000
knapsack_objects <-data.frame( w=sample(1:4000, size = n, replace = TRUE),  
                               v=runif(n = n, 0, 10000))
dynamic_knapsack(x = knapsack_objects[1:20,], W = 2000)
dynamic_knapsack(x = knapsack_objects[1:10,], W = 2000)
}
}
\references{
https://en.wikipedia.org/wiki/Knapsack problem#0.2F1 knapsack problem
}
\author{
Vinod kumar Dasari and Gowtham Kukkemane Mahalingabhat
}
