% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dynamic_programming.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Dynamic programming knapsack}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{A data.frame containing with variables \code{w} and \code{v}}

\item{W}{A numeric value stating the maximum allowed weight of selected objects}
}
\value{
A list containg the best value and the elements selected
}
\description{
A dynamic programming approach for solving knapsack problems.
}
\examples{
n <- 2000
knapsack_objects <- data.frame(w=sample(1:4000, size = n, replace = TRUE), v=runif(n = n, 0, 10000))
knapsack_dynamic(x = knapsack_objects[1:8,], W = 3500)
knapsack_dynamic(x = knapsack_objects[1:12,], W = 3500)
knapsack_dynamic(x = knapsack_objects[1:8,], W = 2000)
knapsack_dynamic(x = knapsack_objects[1:12,], W = 2000)
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack problem#0.2F1 knapsack problem}
}
