% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{knapsack_dynamic function}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{A data.frame with two variables v (values) and w (weights)}

\item{W}{The maximum knapsack size}
}
\value{
A list of the maximum value in the knapsack and the elements choosen to be in the knapsack
}
\description{
knapsack_dynamic function
}
