% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Dynamic programming search for knapsack problem}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{An object of data.frame with 2 columns with colnames w and v.}

\item{W}{A single positive value.}
}
\value{
The function returns a list containing the maximum knapsack value and which elements.
}
\description{
Dynamic programming search for knapsack problem
}
\examples{
set.seed(42)
 n <- 2000
knapsack_objects <-
 data.frame(
   w=sample(1:4000, size = n, replace = TRUE),
   v=runif(n = n, 0, 10000)
 ) 
results <- knapsack_dynamic(x = knapsack_objects[1:8,], W=3500)
}
