% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Dynamic.}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{data.frame}

\item{W}{integer}
}
\value{
list of ideal value and elements that solves the problem.
}
\description{
Solving Knapsack problem with dynamic programming.
}
