% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Dynamic knapsack}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{Knapsack data frame}

\item{W}{Knapsack size}
}
\value{
elements and the sum of the corresponding values
}
\description{
Dynamic knapsack
}
