% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{knapsack_dynamic}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{Dataframe containing weight and value of objects}

\item{W}{maximum weight allowed in knapsack}
}
\value{
List with the optimal value and the respective elements,
that are included in the solution.
}
\description{
Solver for the knapsack problem based on dynamic programming.
}
\examples{
knapsack_objects <- get_knapsack_objects(2000)

knapsack_dynamic(x = knapsack_objects[1:8, ], W = 3500)

}
