% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Knapsack problem using dynamic algortihm}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{a data.frame with two variables v and w (values and weights)}

\item{W}{knapsack size (Max weight capacity)}
}
\value{
the maximum values and elements to be carried by knapsack
}
\description{
Knapsack problem using dynamic algortihm
}
\examples{
knapsack_dynamic(x = knapsack_objects[1:8,], W = 3500)
knapsack_dynamic(x = knapsack_objects[1:12,], W = 2000)


}
