% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Dynammic Programming for Knapsack Problem}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{}

\item{W}{}
}
\value{
It returns the maximum value obtained and selected items.
}
\description{
The algorithm solves the knapsack problem using Dynammic Programming approach. 
Its computational complexity is O(Wn).
}
\examples{
x <- data.frame(v = c(10, 5, 15, 7, 6, 18), w = c(2, 3, 5, 7, 1, 4))
W <- 15
knapsack_dynamic(x,W)
}
\seealso{
[Knapsack Problem](https://en.wikipedia.org/wiki/Knapsack_problem#0.2F1_knapsack_problem)
}
