% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Dynamic Programming for Knapsack Problem}
\usage{
knapsack_dynamic(x, W, fast = FALSE)
}
\arguments{
\item{x}{a data.frame with two variables v (Value) and w (Weight) and returns the maximum knapsack value and which elements (rows in the data.frame).}

\item{W}{is the knapsack size.}

\item{fast}{An optional logical value to use rcpp (default is FALSE)}
}
\value{
It returns the maximum value obtained and selected items.
}
\description{
The algorithm solves the knapsack problem using Dynamic Programming approach. 
Its computational complexity is O(Wn).
}
\examples{
x <- data.frame(w = c(2, 3, 5, 7, 1, 4), v = c(10, 5, 15, 7, 6, 18))
W <- 15
knapsack_dynamic(x,W)
knapsack_dynamic(x,W, fast = TRUE)
}
\seealso{
[Knapsack Problem](https://en.wikipedia.org/wiki/Knapsack_problem#0.2F1_knapsack_problem)
}
