% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Dynamic Programming}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{x \code{data.frame} containing objects with weights \code{w} and values \code{v} as columns.}

\item{W}{max weight (discrete) of the knapsack}
}
\value{
a list with max value and corresponding elements
}
\description{
Dynamic Programming
}
\examples{
data(knapsack_objects)
knapsack_dynamic(x = knapsack_objects[1:8,], W = 3500)
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem#Dynamic_programming_in-advance_algorithm}
}
