% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Dynamic Programming}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{}

\item{W}{}
}
\value{
a list with max value and corresponding elements
}
\description{
Dynamic Programming
}
\examples{
knapsack_objects <-
data.frame(
  v = c(10,40,30,50),
  w = c(5,4,6,3)
)
W <- 10

knapsack_dynamic((knapsack_objects, W)
}
