% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_algos.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Dynamic programming knapsack solver}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{data.frame with columns w, v (weights should be integers or coercible)}

\item{W}{numeric capacity}
}
\value{
list(value = numeric, elements = integer vector of selected original indices)
}
\description{
Dynamic programming knapsack solver
}
\examples{
x <- data.frame(w = c(2,3,4), v = c(3,4,5)); knapsack_dynamic(x, 5)
}
