% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Solving the knapsack problem using dynamic programming}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{an data frame with a number of items.}

\item{W}{the limited weight the knapsack can handle.}
}
\value{
A list of the maximated value with corresponding elements that fits in the bag.
}
\description{
The function solves the knapsack problem , \url{https://en.wikipedia.org/wiki/Knapsack_problem}
using dynamtic programming. The function iterates over all possible values of W.
}
\examples{
knapsack_objects <-data.frame(w=sample(1:4000, size = n, replace = TRUE),v=runif(n = n, 0, 10000))
knapsack_dynamic(x=knapsack_objects[1:500,] W=2000)
}

