% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Solving the knapsack problem using dynamic programming}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{an data frame with a number of items.}

\item{W}{the limited weight the knapsack can handle.}
}
\value{
A list of the maximated value with corresponding elements that fits in the bag.
}
\description{
The function solves the knapsack problem using dynamic programming. The function iterates over all possible values of W. It returns the same value as \code{brute_force_knapsack}.
}
\examples{
n<- 2000
knapsack_objects <-data.frame(w=sample(1:4000, size = n, replace = TRUE),v=runif(n = n, 0, 10000))
knapsack_dynamic(x=knapsack_objects[1:8,], W=2000)
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem#0.2F1_knapsack_problem}
}

