% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_objects.R
\docType{data}
\name{knapsack_objects}
\alias{knapsack_objects}
\title{A Sample Set of Knapsack Objects}
\format{
A data.frame with 2000 observations of two variables: 'w' (weight)
and v' (value)
}
\usage{
knapsack_objects
}
\description{
A set of 2000 generated objects for use with the \emph{knapsack}
package. Weights are randomly assigned between 1 and 4000 and values are
randomly assigned between 1 and 10000
}
\keyword{datasets}
