% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Knapsack_objects.R
\docType{data}
\name{knapsack_objects}
\alias{knapsack_objects}
\title{Knapsack_objects}
\format{An object of class \code{data.frame} with 1000000 rows and 2 columns.}
\usage{
knapsack_objects
}
\description{
Creating a data set
}
\keyword{datasets}
