% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sac_a_dos_glouton.R
\name{sac_a_dos_glouton}
\alias{sac_a_dos_glouton}
\title{Sac a dos : Algorithme Glouton}
\usage{
sac_a_dos_glouton(poids, valeurs, poids_max)
}
\arguments{
\item{poids}{vecteur contenant les poids des objets}

\item{valeurs}{vecteur contenant les valeurs (numériques) des objets}

\item{poids_max}{représente le poids maximal autorisé dans le sac}
}
\value{
Une solution approchée
}
\description{
Algorithme prenant le meilleur rapport valeur/poids
}
